pyproject.toml
[project]
name = "green-agent-template"
version = "0.1.0"
description = "DeFi Risk Auditor Green Agent for AgentBeats"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "boto3>=1.42.30",
    "docker>=7.1.0",
    "fastapi>=0.124.4",
    "httpx[test]>=0.28.1",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "pytest-asyncio[test]>=1.3.0",
    "pytest[test]>=9.0.2",
    "python-dotenv>=1.2.1",
    "snowflake-connector-python>=4.2.0",
    "uvicorn>=0.38.0",
    "web3>=7.14.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"